refactor project: move projectadapter to common package
